+++ x/upgrade/types/upgrade_config.go
@@ -10,8 +10,11 @@ const (
 	// EnablePublicDelegationUpgrade is the upgrade name for enabling public delegation
 	EnablePublicDelegationUpgrade = types.EnablePublicDelegationUpgrade
 
-	// Nagqu is the upgrade name for following features:
+	// Nagqu is the upgrade name for Nagqu upgrade
 	Nagqu = types.Nagqu
+
+	// Pampas is the upgrade name for Pampas upgrade
+	Pampas = types.Pampas
 )
 
 // The default upgrade config for networks
@@ -28,6 +31,10 @@ var (
 		Name:   Nagqu,
 		Height: 471350,
 		Info:   "Nagqu hardfork",
+	}).SetPlan(&Plan{
+		Name:   Pampas,
+		Height: 2427233,
+		Info:   "Pampas hardfork",
 	})
 )